Warn when `uv python upgrade` tries to install over a non-managed `bin` installation (#14084)
`uv python upgrade` had originally had a `--force` option, but we
thought it probably didn't make sense. However, switching `uv python
upgrade` behind the preview flag revealed a bad interaction: upgrading a
minor version when a non-managed interpreter of that version exists in
the `bin` directory displays an error recommending the use of `--force`.

This PR only warns if this case arises with `uv python upgrade`, noting
that install `--force` can be used. It also includes a test for the case
described above.