(f)lock during `uv run`
This is very similar to the locking added for `uv sync`, `uv add`, and
`uv remove` in https://github.com/astral-sh/uv/pull/13869. Improving our
(f)locking in general is tracked in
https://github.com/astral-sh/uv/issues/13883.

This is a minimal change that un-breaks parallel invocations of `uv
run`. It's too easy to confuse locking as in "lockfile generation" and
locking as in "file locking to prevent race conditions", so we probably
need a broader overhaul of how these things are named. It's also too
easy to miss a callsite here, or to accidentally retain one of these
locks across `run_to_completion`, so ideally we could move this file
locking down to a lower-level function that's shared by all these
callers, possibly `pip::operations::install`. I want to get this minimal
change through testing and review before I tackle either of those.